requirements/base.txt
Django==1.6.5
bpython==0.12
django-braces==1.2.2
django-model-utils==1.5.0
logutils==0.3.3
South==0.8.2


djangorestframework
markdown
django-filter

Pillow
django-mailgun

requests
protobuf
